src/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
python-dateutil==2.8.2
pytz==2023.3
pydantic==1.10.13
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
requests==2.31.0
